Be able to be logged as a Tuleap user.
This is a part of [story #14018](https://tuleap.net/plugins/tracker/?aid=14018) have a central management of users and groups using oauth

Now, when you are logged, you should see the Tuleap username of the user used to login in Tuleap.
At the first login of the user in Jenkins, the user will be automaticaly created.
The user information are retrieved thanks to the /userinfo endpoint and access token.