Remove all the Tuleap api call and rewire this plugin with tuleap-api plugin
The main goal of this commit is to make easier the functionnal test which depends of tuleap-api plugin.
The Tuleap api calls are now handled by tuleap-api-plugin.

/!\ IMPORTANT NOTE /!\
The current release version of tuleap-api plugin does not contain the implementation of the needed queries yet.
You have to install manually the tuleap-api.